examples/python/requirements.txt
hvac>=2.1.0
pyyaml>=6.0
requests>=2.31.0
